added exception - User not found

+++ src/main/java/com/userservice/Service/UserService.java
@@ -1,6 +1,7 @@
 package com.userservice.Service;
 
 
+import com.userservice.Exception.UserNotFoundException;
 import com.userservice.Model.User;
 import com.userservice.Model.UserDTO;
 import com.userservice.Repository.UserRepo;
@@ -28,6 +29,9 @@ public class UserService {
         }
         Pageable firstPage = PageRequest.of(page-1, pageSize);
         List<User> allUsers=  userRepo.findAll(firstPage).toList();
+        if(allUsers.isEmpty()){
+            throw new UserNotFoundException("User Not Found");
+        }
         List<UserDTO> allUsersDTO=new ArrayList<>();
         for(User user:allUsers){
             UserDTO userDTO=new UserDTO(user.getUserID(),user.getFirstName(),user.getMiddleName(),
@@ -42,20 +46,28 @@ public class UserService {
     }
 
     public UserDTO findByID(String userId){
-        User user=this.userRepo.findById(userId).get();
-        UserDTO userDTO=new UserDTO();
-        userDTO.setUserID(user.getUserID());
-        userDTO.setFirstName(user.getFirstName());
-        userDTO.setLastName(user.getLastName());
-        userDTO.setMiddleName(user.getMiddleName());
-        userDTO.setPhoneNumber(user.getPhoneNumber());
-        userDTO.setEmail(user.getEmail());
-        userDTO.setAddress(user.getAddress());
-        userDTO.setDateOfBirth(user.getDateOfBirth());
-        userDTO.setEmployeeNumber(user.getEmployeeNumber());
-        userDTO.setBloodGroup(user.getBloodGroup());
-        userDTO.setGender(user.getGender());
-        return  userDTO;
+
+        try{
+            User user=this.userRepo.findById(userId).get();
+            UserDTO userDTO=new UserDTO();
+            userDTO.setUserID(user.getUserID());
+            userDTO.setFirstName(user.getFirstName());
+            userDTO.setLastName(user.getLastName());
+            userDTO.setMiddleName(user.getMiddleName());
+            userDTO.setPhoneNumber(user.getPhoneNumber());
+            userDTO.setEmail(user.getEmail());
+            userDTO.setAddress(user.getAddress());
+            userDTO.setDateOfBirth(user.getDateOfBirth());
+            userDTO.setEmployeeNumber(user.getEmployeeNumber());
+            userDTO.setBloodGroup(user.getBloodGroup());
+            userDTO.setGender(user.getGender());
+            return  userDTO;
+        }
+
+        catch(Exception e){
+            throw new UserNotFoundException("User Not Found");
+        }
+
 
     }
 
@@ -72,14 +84,21 @@ public class UserService {
             return this.userRepo.save(user);
         }
         else{
-            throw new Exception("ID doesnot Exist");
+            throw new UserNotFoundException("User Not Found");
         }
 
     }
 
             public String deleteUserById(String userId){
-                userRepo.deleteById(userId);
-                return "User  Successfully Deleted";
+                if(userRepo.findById(userId).isPresent()){
+
+                    userRepo.deleteById(userId);
+                    return "User  Successfully Deleted";
+                }
+                else{
+                    throw new UserNotFoundException("User Not Found");
+                }
+
             }
